+++ src/main/java/com/x29naybla/bloom_and_doom/common/entity/goal/PlantShootGoal.java
@@ -20,13 +20,13 @@ public class PlantShootGoal extends Goal {
     private final float attackRadiusSqr;
 
     private int shootTimer;
-    private final int timerCap;
+    private final float timerCap;
 
-    public PlantShootGoal(ShootingPlant shootingPlant, int timeInSeconds, int attackInterval, float attackRadius){
+    public PlantShootGoal(ShootingPlant shootingPlant, float timeInSeconds, int attackInterval, float attackRadius){
         this(shootingPlant, timeInSeconds, attackInterval, attackInterval, attackRadius);
     }
 
-    public PlantShootGoal(ShootingPlant shootingPlant, int timeInSeconds, int attackIntervalMin, int attackIntervalMax, float attackRadius){
+    public PlantShootGoal(ShootingPlant shootingPlant, float timeInSeconds, int attackIntervalMin, int attackIntervalMax, float attackRadius){
         this.attackTime = -1;
         if (shootingPlant == null) {
             throw new IllegalArgumentException("PlantShootGoal requires Mob extends ShootingPlant");
